Updated duplicate job logic
Changed the logic when dealing with duplicate jobs for a client so that we now start killing duplicates when we have a long or slow running one in the queue backing them up. The Director will now priortize more important jobs (Full over Incremental), killing currently running ones if nessessary, but if they are of the same type, then we'll keep the running one (if found) or the newest (when both jobs are simply queued).

Also we will now allow higher priority jobs to run when lower priority ones are still running (providing we haven't reached our concurrent limits), instead of waiting. However, this doesn't allow lower priorty jobs to run at the same time as higher.